requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.21
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0
